feat: Add .gitignore file to project
This commit adds a .gitignore file to the project. The .gitignore file specifies which files and directories should be ignored by Git. In this case, it includes common files and directories such as node_modules, __pycache__, .vscode, .task_creation, .idea, and db.sqlite3.

These changes enhance the project by preventing unnecessary files and directories from being tracked by Git.

This commit is a part of the ongoing development of the project.