feat: Add manage.py file
This commit adds the  file to the project. The  file is a command-line utility for administrative tasks in Django. It sets the  environment variable and allows users to run administrative tasks using the  function.

These changes enhance the project by providing a convenient way to manage and run administrative tasks.

This commit is a part of the ongoing development of the project.