feat: Add Tailwind CSS configuration file
This commit adds the  file to the project. The  file is used to configure the Tailwind CSS framework. It specifies the content that should be processed by Tailwind CSS, as well as any customizations to the theme.

These changes enhance the project by allowing the use of Tailwind CSS classes and customizations throughout the project.

This commit is a part of the ongoing development of the project.