feat: Add login page template
This commit adds a new login.html template to the registration folder. The template includes the necessary HTML, CSS, and JavaScript code for a login page. This addition enables users to log in to the Task Creation App.

Note: This commit message follows the established convention of using a prefix to indicate the type of change.