chore: Update dashboard link in base template and rename index.html to dashboard.html
This commit updates the base template (_base.html) to use the correct URL for the dashboard link. The link was previously pointing to the 'index' URL, but it has been renamed to 'dashboard'. Additionally, the commit renames the index.html file to dashboard.html to reflect the change in functionality.

Note: This commit message follows the established convention of using a prefix to indicate the type of change.