feat: Add CustomUserCreationForm to handle user registration
This commit adds a new form, CustomUserCreationForm, to handle user registration in the Task Creation App. The form extends the UserCreationForm provided by Django and includes additional fields for email, first name, and last name. The save method is overridden to save the user with the provided email, first name, and last name. This form will be used in the signup view to create new user accounts.

Note: This commit message follows the established convention of using a prefix to indicate the type of change.