chore: Add post_save signal to create auth token for new users
This commit adds a post_save signal to the models.py file in order to automatically create an authentication token for new users. The signal is triggered when a new user is created and uses the rest_framework.authtoken app to generate a token associated with the user. This functionality ensures that new users have a valid token for authentication purposes.

Note: This commit message follows the established convention of using a prefix to indicate the type of change.