chore: Update assigned_to dropdown in dashboard.html to use user's username
This commit modifies the assigned_to dropdown in the dashboard.html template to dynamically display the username of the currently logged-in user. Instead of hardcoding the options, the value is now set to \{{user.username}}\. This change ensures that the dropdown accurately reflects the user's username and improves the user experience.

Note: This commit message follows the established convention of using a prefix to indicate the type of change.