chore: Add drf-yasg to INSTALLED_APPS
This commit adds the \drf-yasg\ package to the INSTALLED_APPS list in the settings.py file. By including this package, we enable the generation of Swagger documentation for the Task Creation App's API endpoints. This change improves the developer experience by providing an interactive API documentation tool.

Note: This commit message follows the established convention of using a prefix to indicate the type of change.